refactor(verification): shifted over processing admin sms count to verification module
