+++ src/app/modules/form/admin-form/admin-form.service.ts
@@ -9,7 +9,6 @@ import {
   MAX_UPLOAD_FILE_SIZE,
   VALID_UPLOAD_FILE_TYPES,
 } from '../../../../shared/constants'
-import { SMS_WARNING_TIERS } from '../../../../shared/util/verification'
 import {
   BasicField,
   FormLogoState,
@@ -41,11 +40,6 @@ import { isVerifiableMobileField } from '../../../../types/field/utils/guards'
 import { aws as AwsConfig } from '../../../config/config'
 import { createLoggerWithLabel } from '../../../config/logger'
 import getFormModel from '../../../models/form.server.model'
-import {
-  MailGenerationError,
-  MailSendError,
-} from '../../../services/mail/mail.errors'
-import MailService from '../../../services/mail/mail.service'
 import * as SmsService from '../../../services/sms/sms.service'
 import { dotifyObject } from '../../../utils/dotify-object'
 import {
@@ -1142,63 +1136,4 @@ const isMobileFieldUpdateAllowed = (
         : okAsync(form)
     },
   )
-}
-
-/**
- * Checks the number of free smses sent by the admin of the form and deactivates verification or sends mail as required
- * @param form The form whose admin's sms counts needs to be checked
- * @returns ok(true) when the verification has been deactivated successfully or no action is required
- * @returns err(MailGenerationError) when an error occurred on creating the HTML template for the email
- * @returns err(MailSendError) when an error occurred on sending the email
- * @returns err(PossibleDatabaseError) when an error occurred while retrieving the counts from the database
- */
-export const checkFreeSmsSentByAdminAndDeactivateVerification = (
-  form: IPopulatedForm,
-): ResultAsync<
-  true,
-  MailGenerationError | MailSendError | PossibleDatabaseError
-> => {
-  if (isFormOnboarded(form)) {
-    return okAsync(true)
-  }
-
-  // Convert to string because it's typed as any
-  const formAdminId = String(form.admin._id)
-
-  return SmsService.retrieveFreeSmsCounts(formAdminId).andThen((freeSmsSent) =>
-    checkSmsCountAndPerformAction(form, freeSmsSent),
-  )
-}
-
-/**
- * Checks the number of free smses sent by the admin of a form and performs the appropriate action
- * @param form The form whose admin's sms counts needs to be checked
- * @returns ok(true) when the action has been performed successfully
- * @returns err(MailGenerationError) when an error occurred on creating the HTML template for the email
- * @returns err(MailSendError) when an error occurred on sending the email
- * @returns err(PossibleDatabaseError) when an error occurred while retrieving the counts from the database
- */
-const checkSmsCountAndPerformAction = (
-  form: Pick<IPopulatedForm, 'admin' | 'title' | '_id' | 'permissionList'>,
-  freeSmsSent: number,
-): ResultAsync<
-  true,
-  MailGenerationError | MailSendError | PossibleDatabaseError
-> => {
-  // Convert to string because it's typed as any
-  const formAdminId = String(form.admin._id)
-
-  // NOTE: Because the admin has exceeded their allowable limit of free sms,
-  // the sms verifications for their forms also need to be disabled.
-  if (hasAdminExceededFreeSmsLimit(freeSmsSent)) {
-    return MailService.sendSmsVerificationDisabledEmail(form).andThen(() =>
-      disableSmsVerificationsForUser(formAdminId),
-    )
-  }
-
-  if (freeSmsSent in SMS_WARNING_TIERS) {
-    return MailService.sendSmsVerificationWarningEmail(form, freeSmsSent)
-  }
-
-  return okAsync(true)
 }
